requirements.txt
websockets>=12.0
TikTokLive>=6.0.0
flask>=3.0.0
requests>=2.31.0
gunicorn>=22.0.0
psycopg2-binary>=2.9.9
PyJWT>=2.8.0
bcrypt==4.0.1
python-dotenv>=1.0.0
